+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## 0.6.7
+
+
 ## 0.6.6
 
 - remove `@SimpleTask` Class